Added LogAndReg.py via upload
LogAndReg.py is a python file where you can login and register using a username/password combination.  The usernames and passwords are stored on files that are written to the users system and there is no database implemented.  There are also little to no security measures.  The users and passwords are stored in plain text files.